Adicionado - Sigano.py (Modificado)

+++ Gustavo Guanabara/Aula 10 - Condições/Sigano.py
@@ -3,15 +3,15 @@ adv = float(input("Você está preso em um Genjutsu,\nPara sair você terá que 
 if adv>5 or adv<0:
     print("Digito fora do range!!")
 else:
-    cpu = float(random.choice(range(0,6)))
+    cpu = int(random.choice(range(0,6)))
     while True:
         if cpu == adv:
             print("KRLLL !! Tu é bom mesmo !!")
             break
         print("\n\nERRRROOOOOOUUU!")
-        cpu = float(random.choice(range(0,6)))
+        cpu = int(random.choice(range(0,6)))
         print("\nPensei no número {}\n\nTente novamente, loser !!".format(int(cpu)))
-        adv = float(input("Você está preso em um Genjutsu,\nPara sair você terá que acertar qual número estou pensando\nDigite um número: "))
+        adv = int(input("Você está preso em um Genjutsu,\nPara sair você terá que acertar qual número estou pensando\nDigite um número: "))
         if adv>5 or adv<0:
             print("Digito fora do range!!")
             break
